backend/requirements.txt
# RobustnessLab Backend Dependencies
# ===================================

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0

# Data Validation
pydantic>=2.0.0

# OCR Engines (opcional - instale conforme necessário)
# pytesseract>=0.3.10  # Requer Tesseract instalado no sistema
# easyocr>=1.7.0       # Deep learning based OCR
# paddleocr>=2.7.0     # Alta precisão OCR

# Development
python-dotenv>=1.0.0
